+++ packages/brick-kit/src/bindListeners.spec.ts
@@ -11,6 +11,16 @@ import * as runtime from "./core/Runtime";
 
 jest.mock("./history");
 
+// Mock a custom element of `any-provider`.
+customElements.define(
+  "any-provider",
+  class Tmp extends HTMLElement {
+    resolve(): string {
+      return "resolved";
+    }
+  }
+);
+
 const mockHistory = {
   push: jest.fn(),
   replace: jest.fn(),
@@ -44,6 +54,20 @@ jest.spyOn(runtime, "_internalApiGetCurrentContext").mockReturnValue({
           alias: "segue-target-b",
         },
       ],
+      [
+        "mock-alias-a",
+        {
+          path: "/mock/alias/a",
+          alias: "mock-alias-a",
+        },
+      ],
+      [
+        "mock-alias-b",
+        {
+          path: "/mock/alias/b/:type",
+          alias: "mock-alias-b",
+        },
+      ],
     ]),
   },
   segues: {
@@ -56,6 +80,17 @@ jest.spyOn(runtime, "_internalApiGetCurrentContext").mockReturnValue({
   },
   storyboardContext,
 } as any);
+
+const anyProvider = document.createElement("any-provider");
+jest.spyOn(runtime, "_internalApiGetProviderBrick").mockImplementation(
+  async (provider: string): Promise<HTMLElement> => {
+    await Promise.resolve();
+    if (provider === "any-provider") {
+      return anyProvider;
+    }
+    throw new Error(`Provider not defined: "${provider}".`);
+  }
+);
 
 describe("isBuiltinHandler", () => {
   const cases: [BrickEventHandler, boolean][] = [
@@ -216,6 +251,11 @@ describe("bindListeners", () => {
           action: "segue.replace",
           args: ["testSegueIdB", { id: "${EVENT.detail}" }],
         },
+        { action: "alias.push", args: ["mock-alias-a"] },
+        {
+          action: "alias.replace",
+          args: ["mock-alias-b", { type: "${EVENT.detail}" }],
+        },
         { action: "event.preventDefault" },
         { action: "console.log" },
         { action: "console.info" },
@@ -316,6 +356,23 @@ describe("bindListeners", () => {
           target: "#target-elem",
           properties: { someProperty: "${EVENT.detail}" },
         },
+        {
+          useProvider: "any-provider",
+          args: ["for", "${EVENT.detail}"],
+          callback: {
+            success: {
+              action: "console.log",
+            },
+          },
+        },
+        {
+          useProvider: "not-defined-provider",
+          callback: {
+            success: {
+              action: "console.log",
+            },
+          },
+        },
       ],
       key3: { action: "not.existed" },
       key4: {},
@@ -361,6 +418,7 @@ describe("bindListeners", () => {
     const history = mockHistory;
     expect(history.push).toHaveBeenNthCalledWith(1, "for-good");
     expect(history.push).toHaveBeenNthCalledWith(2, "/segue-target-a");
+    expect(history.push).toHaveBeenNthCalledWith(3, "/mock/alias/a");
     expect(history.pushQuery).toBeCalledWith(
       {
         q: "123",
@@ -381,6 +439,10 @@ describe("bindListeners", () => {
       2,
       "/segue-target-b/for-good"
     );
+    expect(history.replace).toHaveBeenNthCalledWith(
+      3,
+      "/mock/alias/b/for-good"
+    );
     expect(history.replaceQuery).toBeCalledWith({
       page: 1,
     });
@@ -396,19 +458,23 @@ describe("bindListeners", () => {
 
     expect(spyOnPreventDefault).toBeCalled();
 
-    expect(console.log).toBeCalledTimes(2);
+    expect(console.log).toBeCalledTimes(3);
     expect(console.log).toHaveBeenNthCalledWith(1, event1);
     expect((console.log as jest.Mock).mock.calls[1][0].type).toBe(
       "callback.success"
     );
     expect((console.log as jest.Mock).mock.calls[1][0].detail).toBe("yes");
+    expect((console.log as jest.Mock).mock.calls[2][0].type).toBe(
+      "callback.success"
+    );
+    expect((console.log as jest.Mock).mock.calls[2][0].detail).toBe("resolved");
     expect(console.info).toBeCalledTimes(2);
     expect(console.info).toBeCalledWith(event1);
     expect((console.info as jest.Mock).mock.calls[1][0].type).toBe(
       "callback.finally"
     );
     expect(console.warn).toBeCalledWith("specified args for console.warn");
-    expect(console.error).toBeCalledTimes(6);
+    expect(console.error).toBeCalledTimes(7);
     expect(console.error).toHaveBeenNthCalledWith(
       1,
       "specified args for console.error"
@@ -424,10 +490,14 @@ describe("bindListeners", () => {
       "target not found:",
       "#not-existed"
     );
-    expect((console.error as jest.Mock).mock.calls[5][0].type).toBe(
+    expect(console.error).toHaveBeenNthCalledWith(
+      6,
+      'Error: Provider not defined: "not-defined-provider".'
+    );
+    expect((console.error as jest.Mock).mock.calls[6][0].type).toBe(
       "callback.error"
     );
-    expect((console.error as jest.Mock).mock.calls[5][0].detail).toBe("oops");
+    expect((console.error as jest.Mock).mock.calls[6][0].detail).toBe("oops");
     expect((sourceElem as any).forGood).toHaveBeenNthCalledWith(
       1,
       "target is _self"
@@ -571,6 +641,7 @@ describe("bindListeners", () => {
         if: "<% !EVENT.detail.rejected %>",
       },
       { action: "segue.push", if: "<% !EVENT.detail.rejected %>" },
+      { action: "alias.push", if: "<% !EVENT.detail.rejected %>" },
       { action: "context.assign", if: "<% !EVENT.detail.rejected %>" },
       { action: "event.preventDefault", if: "<% !EVENT.detail.rejected %>" },
       { action: "console.log", if: "<% !EVENT.detail.rejected %>" },