+++ packages/brick-types/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.46.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.45.1...@easyops/brick-types@1.46.0) (2020-07-17)
+
+### Features
+
+- **useResolves:** support useProvider ([d31e46b](https://git.easyops.local/anyclouds/next-core/commits/d31e46b))
+
+## [1.45.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.45.0...@easyops/brick-types@1.45.1) (2020-07-14)
+
+### Bug Fixes
+
+- **context:** define brick property as context ([2ada956](https://git.easyops.local/anyclouds/next-core/commits/2ada956))
+
+# [1.45.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.44.0...@easyops/brick-types@1.45.0) (2020-07-13)
+
+### Features
+
+- add login from refs ([60cdc51](https://git.easyops.local/anyclouds/next-core/commits/60cdc51))
+
+# [1.44.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.43.1...@easyops/brick-types@1.44.0) (2020-07-08)
+
+### Features
+
+- route by alias ([67448fe](https://git.easyops.local/anyclouds/next-core/commits/67448fe))
+
 ## [1.43.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.43.0...@easyops/brick-types@1.43.1) (2020-07-06)
 
 **Note:** Version bump only for package @easyops/brick-types